Pesquisa.

+++ src/main/java/br/com/localizaja/dto/BuscaEstabelecimentoDTO.java
@@ -1,5 +1,7 @@
 package br.com.localizaja.dto;
 
+import java.math.BigDecimal;
+
 /**
  *
  * @author adriana
@@ -12,9 +14,9 @@ public class BuscaEstabelecimentoDTO {
 
     private String enderecoOrigem;
 
-    private Double latitude;
+    private BigDecimal latitude;
 
-    private Double longitude;
+    private BigDecimal longitude;
 
     public String getSeguimento() {
         return seguimento;
@@ -40,19 +42,19 @@ public class BuscaEstabelecimentoDTO {
         this.enderecoOrigem = enderecoOrigem;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(Double latitude) {
+    public void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(Double longitude) {
+    public void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
